split route about widget-list and widget-setting
