+++ app/components/WidgetList/components/ToggleButtonWithLabel/ToggleButtonWithLabel.scss
