+++ app/containers/WidgetListContainer.jsx
@@ -1,0 +1,34 @@
+import { connect } from 'react-redux';
+import WidgetList from 'components/WidgetList';
+import GNBWrapper from 'components/GlobalNavigationBar/GNBWrapper';
+import {
+  widgetListInfoStore,
+  widgetListSelect,
+  widgetInfoUpdateWithIPC,
+} from 'store/widget/actions';
+import {
+  modalOpen,
+  modalClose,
+} from 'store/modal/actions';
+
+import {
+  selectedIdSelector,
+  getSelectedWidget,
+  getWidgetInfo,
+} from 'store/widget/selectors';
+
+const mapStateToProps = state => ({
+  list: getWidgetInfo(state),
+  selectedId: selectedIdSelector(state),
+  selectedWidget: getSelectedWidget(state),
+});
+
+const mapDispatchToProps = {
+  onModalClose: modalClose,
+  onModalOpen: modalOpen,
+  onStoreWidgetInfo: widgetListInfoStore,
+  onSelectItem: widgetListSelect,
+  onUpdateInfoWithIPC: widgetInfoUpdateWithIPC,
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(GNBWrapper(WidgetList));
